02-bot/requirements.txt
tqdm
python-dotenv
openai
Flask
gunicorn
python-telegram-bot
pydantic-settings
pydantic